✨ Refactor feedback and error handling logic
- Renamed `WeatherApp` to `WeatherFitApp` in `lib/weather_app.dart` (now `lib/app/weather_fit_app.dart`).
- Extracted feedback and error handling logic from `WeatherFitApp` into a new `SettingsStateListenerContent` widget.
  - `lib/app/weather_fit_app.dart`:
    - Converted to a `StatelessWidget`.
    - Removed feedback-related state and methods.
    - Added `navigatorKey` to `MaterialApp`.
    - Wrapped `MaterialApp`'s child with `SettingsStateListenerContent` using the `builder` property.
  - `lib/app/settings_state_listener_content.dart`:
    - New `StatefulWidget` to manage feedback UI, automatic feedback submission, loading dialogs, and error notifications.
    - Uses `navigatorKey.currentContext` for showing dialogs and navigation to ensure the correct context is used, especially after asynchronous operations.
- `lib/search/ui/search_page.dart`:
  - Updated error handling for `SearchError` state:
    - On extra small screens (watches), now shows a `showGeneralDialog` with a custom layout for GPS unavailable errors, including "Report issue" and "Close" actions.
    - For other screen sizes, continues to use `ScaffoldMessenger.of(context).showSnackBar`.
  - Ensured `mounted` check before accessing `context` in `_handleReportErrorSnackBarAction`.
- `lib/router/navigator.dart`:
  - Created a new file to hold the global `navigatorKey`.
- `lib/main.dart`:
  - Updated import and usage from `WeatherApp` to `WeatherFitApp`.
- Translations:
  - `assets/i18n/en.json`: Added `gps_unavailable_watch` translation.
  - `assets/i18n/uk.json`: Added `gps_unavailable_watch` translation.
