+++ assets/i18n/uk.json
@@ -111,7 +111,8 @@
     "unable_to_connect_short": "Неможливо підключитися. Спробувати ще?",
     "location_unavailable": "Не вдалося визначити ваше місцезнаходження.",
     "location_unavailable_details": "Дозвіл на доступ до геолокації увімкнено, але застосунку все одно не вдалося визначити вашу позицію. Вибачте за незручності.",
-    "location_unavailable_suggestions": "Ви можете ввести назву іншого міста чи країни, повідомити розробнику про проблему або видалити застосунок."
+    "location_unavailable_suggestions": "Ви можете ввести назву іншого міста чи країни, повідомити розробнику про проблему або видалити застосунок.",
+    "gps_unavailable_watch": "Схоже, що GPS на вашому годиннику недоступний для цього додатка. Ви можете спробувати ввести інше місто чи країну, повідомити про проблему або видалити додаток."
   },
   "settings": {
     "title": "Налаштування",
