+++ lib/main.dart
@@ -5,6 +5,7 @@ import 'package:nominatim_api/nominatim_api.dart';
 import 'package:open_meteo_api/open_meteo_api.dart';
 import 'package:shared_preferences/shared_preferences.dart'
     show SharedPreferences;
+import 'package:weather_fit/app/weather_fit_app.dart';
 import 'package:weather_fit/data/data_sources/local/local_data_source.dart';
 import 'package:weather_fit/data/repositories/location_repository.dart';
 import 'package:weather_fit/data/repositories/outfit_repository.dart';
@@ -13,17 +14,16 @@ import 'package:weather_fit/entities/enums/language.dart';
 import 'package:weather_fit/feedback/feedback_form.dart';
 import 'package:weather_fit/localization/localization_delelegate_getter.dart'
     as locale;
-import 'package:weather_fit/weather_app.dart';
 import 'package:weather_repository/weather_repository.dart';
 
 /// The [main] is the ultimate detail — the lowest-level policy.
 /// It is the initial entry point of the system.
 /// Nothing, other than the operating system, depends on it.
 /// Here you should [injectDependencies].
-/// Think of [main] as a plugin to the [WeatherApp] - a plugin that sets up
+/// Think of [main] as a plugin to the [WeatherFitApp] - a plugin that sets up
 /// the initial conditions and configurations, gathers all the outside
 /// resources, and then hands control over to the high-level policy of the
-/// [WeatherApp].
+/// [WeatherFitApp].
 /// When [main] is released, it has utterly no effect on any of the other
 /// components in the system. They don’t know about [main], and they don’t care
 /// when it changes.
@@ -71,7 +71,7 @@ void main() async {
               );
             },
         theme: FeedbackThemeData(feedbackSheetColor: Colors.grey.shade50),
-        child: WeatherApp(
+        child: WeatherFitApp(
           weatherRepository: WeatherRepository(),
           locationRepository: LocationRepository(
             NominatimApiClient(),
